LayoutTests/fast/css/rgba-custom-text.html

<!DOCTYPE HTML>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style>
   #test {
       color: rgba(0, 0, 0, 0.2);
   }
</style>
<div id="test"></div>
<script>
    description("Tests the CSS custom text implementation for rgba(). Note that WebKit internally stores alpha as an 8 bit value, and that 255 * 0.2 = 51 exactly. So 0.2 is able to be stored with no rounding.");
    shouldBeEqualToString('getComputedStyle(document.getElementById("test"), null).color', "rgba(0, 0, 0, 0.2)");
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
