third_party/WebKit/LayoutTests/fast/css/rgba-custom-text-expected.txt

Tests the CSS custom text implementation for rgba(). Note that WebKit internally stores alpha as an 8 bit value, and that 255 * 0.2 = 51 exactly. So 0.2 is able to be stored with no rounding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("test"), null).color is "rgba(0, 0, 0, 0.2)"
PASS successfullyParsed is true

TEST COMPLETE
